fix: pytest discovery, broken by non-tests in test folder
The run_* files are actually examples, and their presence in the tests/ folder only served to obscure their purpose and confuse the default pytest discovery behavior.

Further, these examples (run_* files) have had the `if __name__ == "__main__"` idiom added.
